requirements.txt
fastapi
httpx
python-dotenv
pydantic-settings
uvicorn[standard]
supabase